add a failing test case for multiple model calls
As noted[1] by @viniciussbs, there's a flaw in this implementation. When
`model()` is called twice on a route due to a query params change, all
watchQuery subscriptions are cancelled, including the brand new one with
the updated query params.

Really we should only be cancelling the old queries that existed prior
to the model reload.

[1]: https://github.com/bgentry/ember-apollo-client/pull/20#discussion_r128326261