only unsubscribe from stale subs, unless exiting
When re-fetching a model (likely due to a change in query parameters),
we only want to end any old ("stale") subscriptions that existed prior
to the new model(s) being loaded. To do this, we mark preexisting
subscriptions as stale when `beforeModel` is called.

Additionally, we need to clear out *all* subscriptions when the route is
being torn down, so we special case this scenario in the
RouteQueryManager.