handle undefined data while still loading
When using a 'cache-and-network' fetchPolicy, Apollo calls the
Observer's `next()` as soon as it tries to load data from the cache.
Often there is no data in the cache, so that's just undefined. This
blows up when trying to get properties from undefined.

We don't want to resolve a promise in this scenario, because data is
still loading, and Ember users expect their promises to resolve when the
data has finished loading. As such, we just bail out in this case and
wait for the eventual call to `next()` when there is actual data.

Fixes #45. Closes #56.