Implement queryManager as a macro
This uses `computed` from `@ember-decorators/object`.

This allow us to use computed as a macro in classic ember objects using `EmberOjbect.extend({ ... })` as well as a decorator when using classes.

In the future, `@ember-decorators/object` would only serve us as a polyfill for older Ember versions as this is becoming the default behavior in Ember per [Decorators RFC](https://github.com/emberjs/rfcs/pull/408)